Moved PercentageWithIcon from _atoms to _molecules since it consists of several elements
